Moved ChunkMaterial to its own file
