+++ src/world/rendering/material.rs
@@ -1,0 +1,166 @@
+use bevy::prelude::*;
+
+use bevy::{
+    ecs::system::lifetimeless::SRes,
+    pbr::MaterialPipeline,
+    reflect::TypeUuid,
+    render::{
+        mesh::MeshVertexAttribute,
+        render_asset::{PrepareAssetError, RenderAsset, RenderAssets},
+        render_resource::{
+            std140::{AsStd140, Std140},
+            BindGroup, BindGroupDescriptor, BindGroupEntry, BindGroupLayoutDescriptor,
+            BindGroupLayoutEntry, BindingResource, BindingType, Buffer, BufferBindingType,
+            BufferInitDescriptor, BufferSize, BufferUsages, SamplerBindingType, ShaderStages,
+            TextureSampleType, TextureViewDimension, VertexFormat,
+        },
+        renderer::RenderDevice,
+    },
+};
+
+#[derive(Component)]
+pub struct ChunkMaterialHandle(pub Handle<ChunkMaterial>);
+
+#[derive(Debug, Clone, TypeUuid)]
+#[uuid = "f690fd1e-d5d8-45ab-8225-97e2a3f056e0"]
+pub struct ChunkMaterial {
+    pub tile_texture_size: f32,
+    pub texture: Handle<Image>,
+}
+
+impl ChunkMaterial {
+    pub const ATTRIBUTE_TILE_COORD_START: MeshVertexAttribute =
+        MeshVertexAttribute::new("TileCoordStart", 66438, VertexFormat::Float32x2);
+}
+
+#[derive(Clone)]
+pub struct GpuChunkMaterial {
+    _buffer: Buffer,
+    bind_group: BindGroup,
+}
+
+impl Material for ChunkMaterial {
+    fn bind_group(
+        material: &<Self as RenderAsset>::PreparedAsset,
+    ) -> &bevy::render::render_resource::BindGroup {
+        &material.bind_group
+    }
+
+    fn bind_group_layout(
+        render_device: &RenderDevice,
+    ) -> bevy::render::render_resource::BindGroupLayout {
+        render_device.create_bind_group_layout(&BindGroupLayoutDescriptor {
+            entries: &[
+                BindGroupLayoutEntry {
+                    binding: 0,
+                    visibility: ShaderStages::FRAGMENT,
+                    ty: BindingType::Texture {
+                        sample_type: TextureSampleType::Float { filterable: true },
+                        view_dimension: TextureViewDimension::D2,
+                        multisampled: false,
+                    },
+                    count: None,
+                },
+                BindGroupLayoutEntry {
+                    binding: 1,
+                    visibility: ShaderStages::FRAGMENT,
+                    ty: BindingType::Sampler(SamplerBindingType::Filtering),
+                    count: None,
+                },
+                BindGroupLayoutEntry {
+                    binding: 2,
+                    visibility: ShaderStages::VERTEX_FRAGMENT,
+                    ty: BindingType::Buffer {
+                        ty: BufferBindingType::Uniform,
+                        has_dynamic_offset: false,
+                        min_binding_size: BufferSize::new(f32::std140_size_static() as u64),
+                    },
+                    count: None,
+                },
+            ],
+            label: None,
+        })
+    }
+
+    fn fragment_shader(asset_server: &AssetServer) -> Option<Handle<Shader>> {
+        Some(asset_server.load("shaders/voxel.wgsl"))
+    }
+
+    fn vertex_shader(asset_server: &AssetServer) -> Option<Handle<Shader>> {
+        Some(asset_server.load("shaders/voxel.wgsl"))
+    }
+
+    fn specialize(
+        _pipeline: &MaterialPipeline<Self>,
+        descriptor: &mut bevy::render::render_resource::RenderPipelineDescriptor,
+        layout: &bevy::render::mesh::MeshVertexBufferLayout,
+    ) -> Result<(), bevy::render::render_resource::SpecializedMeshPipelineError> {
+        let vertex_layout = layout.get_layout(&[
+            Mesh::ATTRIBUTE_POSITION.at_shader_location(0),
+            Mesh::ATTRIBUTE_NORMAL.at_shader_location(1),
+            Mesh::ATTRIBUTE_UV_0.at_shader_location(2),
+            ChunkMaterial::ATTRIBUTE_TILE_COORD_START.at_shader_location(3),
+        ])?;
+        descriptor.vertex.buffers = vec![vertex_layout];
+        Ok(())
+    }
+}
+
+impl RenderAsset for ChunkMaterial {
+    type ExtractedAsset = ChunkMaterial;
+    type PreparedAsset = GpuChunkMaterial;
+    type Param = (
+        SRes<RenderDevice>,
+        SRes<RenderAssets<Image>>,
+        SRes<MaterialPipeline<Self>>,
+    );
+
+    fn extract_asset(&self) -> Self::ExtractedAsset {
+        self.clone()
+    }
+
+    fn prepare_asset(
+        extracted_asset: Self::ExtractedAsset,
+        (render_device, gpu_images, material_pipeline): &mut bevy::ecs::system::SystemParamItem<
+            Self::Param,
+        >,
+    ) -> Result<
+        Self::PreparedAsset,
+        bevy::render::render_asset::PrepareAssetError<Self::ExtractedAsset>,
+    > {
+        let buffer = render_device.create_buffer_with_data(&BufferInitDescriptor {
+            label: None,
+            contents: extracted_asset.tile_texture_size.as_std140().as_bytes(),
+            usage: BufferUsages::UNIFORM | BufferUsages::COPY_DST,
+        });
+
+        let gpu_image = match gpu_images.get(&extracted_asset.texture) {
+            Some(gpu_image) => gpu_image,
+            None => return Err(PrepareAssetError::RetryNextUpdate(extracted_asset)),
+        };
+
+        let bind_group = render_device.create_bind_group(&BindGroupDescriptor {
+            label: None,
+            layout: &material_pipeline.material_layout,
+            entries: &[
+                BindGroupEntry {
+                    binding: 0,
+                    resource: BindingResource::TextureView(&gpu_image.texture_view),
+                },
+                BindGroupEntry {
+                    binding: 1,
+                    resource: BindingResource::Sampler(&gpu_image.sampler),
+                },
+                BindGroupEntry {
+                    binding: 2,
+                    resource: buffer.as_entire_binding(),
+                },
+            ],
+        });
+
+        Ok(GpuChunkMaterial {
+            _buffer: buffer,
+            bind_group,
+        })
+    }
+}
